pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prj2env"
version = "0.1.1"
description = "Transfers dependencies from pyproject.toml to environment.yml"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["."]


[project.scripts]
prj2env = "main:main"
